+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `calfw-toolbar-buttons` to customize buttons on the toolbar.
 - Removed `calfw-default-text-sorter`, `calfw-sorter-start-time` is not set by default.
 - `cfw:event` is saved as text property instead of `cfw:source` when an event is rendered.
+- Added customization `calfw-fchar-period-line`.
 
 ## 2.0
 - Renamed all symbols to start with the package name.